# docs/ai-system/OPENAPI_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: CoinDaily AI System API
  version: 1.0.0
  description: |
    Comprehensive REST API documentation for CoinDaily's AI System.
    
    ## Overview
    The CoinDaily AI System provides a comprehensive suite of AI-powered features for cryptocurrency content management,
    market analysis, social media automation, and more.
    
    ## Base URL
    - Production: `https://api.coindaily.com`
    - Staging: `https://staging-api.coindaily.com`
    - Development: `http://localhost:4000`
    
    ## Authentication
    All endpoints require JWT authentication unless otherwise specified.
    Include the JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - Free tier: 100 requests/minute
    - Premium tier: 500 requests/minute
    - Enterprise tier: 2000 requests/minute
    
    ## Response Format
    All responses follow this structure:
    ```json
    {
      "data": {},
      "error": null,
      "cache": {
        "hit": true,
        "expires_at": "2025-10-20T12:00:00Z"
      }
    }
    ```
    
  contact:
    name: CoinDaily API Support
    email: api@coindaily.com
    url: https://docs.coindaily.com
  license:
    name: Proprietary
    url: https://coindaily.com/license
servers:
  - url: https://api.coindaily.com
    description: Production server
  - url: https://staging-api.coindaily.com
    description: Staging server
  - url: http://localhost:4000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
...
security:
  - bearerAuth: []
paths:
...
  /api/ai/audit/logs:
    get:
      tags:
        - Audit
      summary: Get audit logs
      parameters:
        - name: agent_id
          in: query
          schema:
            type: string
            format: uuid
        - name: event_type
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Audit logs retrieved
  /api/ai/audit/export:
    post:
      tags:
        - Audit
      summary: Generate compliance report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                start_date:
                  type: string
                  format: date-time
                end_date:
                  type: string
                  format: date-time
                format:
                  type: string
                  enum: [json, csv, xml]
                  default: json
      responses:
        '200':
          description: Report generated

  # ==================== COSTS ====================
...